Only compiler help for Solaris on CRAN.
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/bfp@31 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/src/hyperg.cpp
@@ -49,7 +49,7 @@ double logPsi(double b, double c, int n, int p, double R2){
     		- M_LN2
     		+ tauMode
     		- logspace_sub(
-    				log(n - 1) + log1p(-R2) - 2 * log1p((1 - R2) * gMode),
+    				log(static_cast<double>(n - 1)) + log1p(-R2) - 2 * log1p((1 - R2) * gMode),
     				log(n - 1 - pmodel - c) - 2 * log1p(gMode) 
     			);
 //   		- log(
